requirements.txt
discord.py==1.4.1
googletrans==3.0.0